Remove dead code Moved data in test to [DataRow]
